requirements.txt
configparser>=3.5.0
python-telegram-bot>=12.2.0
pySerial>=3.4